class as a separate file
